+++ SpeedyWeather/src/parameterizations/land/snow.jl
@@ -9,18 +9,46 @@ from the diagnosed precipitation, melts once the top soil layer exceeds
 `melting_threshold`, and is capped at `snow_depth_cap` to limit infinite snow/ice accumulation
 over perennial ice caps and glaciers.
 $(TYPEDFIELDS)"""
-@parameterized @kwdef struct SnowModel{NF} <: AbstractSnow
+@parameterized @kwdef struct SnowModel{NF, GridVariable2D} <: AbstractSnow
+    "[OPTION] filename of 'permanent' glacier mask"
+    file::String = "glacier_mask.nc"
+
+    "[OPTION] path to the folder containing the glacier mask"
+    path::String = joinpath("data", "boundary_conditions", file)
+
+    "[OPTION] flag to check for glacier mask in SWA or locally"
+    from_assets::Bool = true
+
+    "[OPTION] variable name in netcdf file"
+    varname::String = "glm"
+
+    "[OPTION] SpeedyWeatherAssets version number"
+    version::VersionNumber = DEFAULT_ASSETS_VERSION
+
+    "[OPTION] Grid the albedo file comes on"
+    FieldType::Type{<:AbstractField} = FullGaussianField
+
     "[OPTION] Temperature threshold for snow melting [K]"
     @param melting_threshold::NF = 275 (bounds = Positive,)
 
     "[OPTION] Permanent snow/ice depth cap in equivalent liquid water depth [m]"
     snow_depth_cap::NF = 10
+
+    "'Permanent' glacier mask"
+    snow_depth::GridVariable2D
 end
 
 Adapt.@adapt_structure SnowModel
 
 # generator function
-SnowModel(SG::SpectralGrid, geometry::LandGeometryOrNothing = nothing; kwargs...) = SnowModel{SG.NF}(; kwargs...)
+# SnowModel(SG::SpectralGrid, geometry::LandGeometryOrNothing = nothing; kwargs...) = SnowModel{SG.NF, SG.GridVariable2D}(; kwargs...)
+
+function SnowModel(SG::SpectralGrid, geometry::LandGeometryOrNothing = nothing; kwargs...)
+    (; NF, GridVariable2D, grid) = SG
+    snow_depth = zeros(GridVariable2D, grid)  # initialize snow depth to zero
+    return SnowModel{NF, GridVariable2D}(; snow_depth, kwargs...)
+end
+
 
 function variables(::SnowModel)
     return (
@@ -34,8 +62,30 @@ end
 initialize!(snow::SnowModel, model::PrimitiveEquation) = nothing
 
 # set initial conditions for snow depth in initial conditions
-initialize!(vars::Variables, snow::SnowModel, model::PrimitiveEquation) =
-    set!(vars.prognostic.land, model.geometry, snow_depth = 0)
+function initialize!(
+        vars::Variables,
+        snow::SnowModel,
+        model::PrimitiveEquation,
+    )
+    glm = get_asset(
+        snow.path,
+        from_assets = snow.from_assets,
+        name = snow.varname,
+        ArrayType = snow.FieldType,
+        FileFormat = NCDataset,
+        version = snow.version,
+    )
+
+    glm = on_architecture(model.architecture, glm)
+    glm *= 10  # Convert from binary mask to snow depth (permanent glaciers are 10m)
+
+    # Interpolate onto the grid
+    snow_depth = snow.snow_depth
+    interpolator = RingGrids.interpolator(snow_depth, glm, NF = Float32)
+    interpolate!(snow_depth, glm, interpolator)
+
+    return set!(vars, model.geometry, snow_depth = glm; namespace = :land)
+end
 
 function timestep!(
         vars::Variables,