+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Learned + dynamic surface albedo [#1013](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1013)
+- Data-driven surface roughness scheme [#959](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/959)
+- Dependency for DomainSets updated to v0.7 and v0.8 [#1045](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1045) [#1047](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1047)
 - Fixed docs broken layout due to `NetCDFOutput` by using @docs instead of @doc [#1098](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1098)
 - Enzyme is a weakdep again instead of dep (this was a mistake) [#1097](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1097)
 - Fixed several typos in the documentation [#1094](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1094)